// Week08/Dawon-Y/mission/src/modules/users/dtos/user.dto.ts
// 요청 DTO
export interface UserSignUpRequest {
  email: string;
  name: string;
  gender: string;
  age: number;
  address?: string;
  specAddress?: string;
  phoneNum: string;
  preferences: number[];
}

// 응답 DTO
export interface UserSignUpResponse {
  email: string;
  name: string;
  preferCategory: string[];
}